refactor(studio): move site-build runtime helpers
